+++ gsa/src/web/pages/reports/detailscontent.js
@@ -295,16 +295,16 @@ const PageContent = ({
   const userTagsCount = userTags.length;
 
   const {
-    applications,
-    closed_cves,
-    cves,
-    errors,
-    hosts,
-    operatingsystems,
-    ports,
-    results,
+    applications = {},
+    closed_cves = {},
+    cves = {},
+    errors = {},
+    hosts = {},
+    operatingsystems = {},
+    ports = {},
+    results = {},
     result_count = {},
-    tlsCertificates,
+    tlsCertificates = {},
     timestamp,
     scan_run_status,
   } = report;
@@ -370,6 +370,7 @@ const PageContent = ({
             createFilterType="result"
             filter={filter}
             filters={filters}
+            isLoading={isLoading || isUpdating}
             onEditClick={onFilterEditClick}
             onError={onError}
             onFilterCreated={onFilterCreated}
@@ -467,7 +468,7 @@ const PageContent = ({
                 </Tab>
               </TabList>
             </TabLayout>
-            {isDefined(results) ? (
+            {isDefined(report) ? (
               <Tabs active={activeTab}>
                 <TabPanels>
                   <TabPanel>
@@ -479,7 +480,7 @@ const PageContent = ({
                   </TabPanel>
                   <TabPanel>
                     <ResultsTab
-                      counts={results.counts}
+                      counts={isDefined(results.counts) ? results.counts : {}}
                       delta={delta}
                       filter={filter}
                       hasTarget={!isContainer}