+++ gsa/src/web/pages/tasks/autodeletereportsgroup.js
@@ -53,7 +53,8 @@ const AutoDeleteReportsGroup = ({
       />
       <Spinner
         type="int"
-        min="1"
+        min="2"
+        max="1200"
         name="auto_delete_data"
         value={autoDeleteData}
         disabled={autoDelete !== AUTO_DELETE_KEEP}