+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [20.04] - unreleased
 
 ### Added
+- Multistep dialog feature, implemented on scanner dialog [#1725] (https://github.com/greenbone/gsa/pull/1725)
 
 ### Changed
 
@@ -20,11 +21,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### [9.0.1] - unreleased
 
 ### Added
-- Multistep dialog feature, implemented on scanner dialog [#1725] (https://github.com/greenbone/gsa/pull/1725)
 
 ### Changed
+- Update node dependencies to latest releases [#1735](https://github.com/greenbone/gsa/pull/1735)
+- If details not defined, then choose first OS from identifiers [#1719](https://github.com/greenbone/gsa/pull/1719)
+- Sorting of SecInfo items [#1717](https://github.com/greenbone/gsa/pull/1717))
+- Disable edit PowerFilter icon if isLoading [#1714](https://github.com/greenbone/gsa/pull/1714)
+- Don't show empty menu section [#1711](https://github.com/greenbone/gsa/pull/1711)
+- Increase clickable area for Logout [#1711](https://github.com/greenbone/gsa/pull/1711)
 
 ### Fixed
+- Fixed allowed range for auto delete keep value [#1729](https://github.com/greenbone/gsa/pull/1729)
+- Fix Created column on Results listpage [#1726](https://github.com/greenbone/gsa/pull/1726)
+- Get delta report needs to send details=1 [#1724](https://github.com/greenbone/gsa/pull/1724)
 - Improve filter handling in report details [#1708](https://github.com/greenbone/gsa/pull/1708)
 - Fix TLS certificate download [#1704](https://github.com/greenbone/gsa/pull/1704)
 - Import additional polyfills for IE11 [#1702](https://github.com/greenbone/gsa/pull/1702)
@@ -38,6 +47,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [9.0.0] - 2019-10-14
 
 ### Added
+- Add statereducer function to Select component: Scrolls to last selected item [#1715](https://github.com/greenbone/gsa/pull/1715)
 - Add loading indicator to svg icon [#1701](https://github.com/greenbone/gsa/pull/1701)
 - Update German Translation [#1689](https://github.com/greenbone/gsa/pull/1689)
 - List NVT of the found CVEs at the report details page [#1673](https://github.com/greenbone/gsa/pull/1673)
@@ -68,6 +78,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   [#1466](https://github.com/greenbone/gsa/pull/1466) [#1467](https://github.com/greenbone/gsa/pull/1467)
 
 ### Changed
+- Tweaked LDAP and RADIUS pages to be more consistent [#1718](https://github.com/greenbone/gsa/pull/1718)
 - Decide whether to default to full and fast scan config (task dialog)[#1671](https://github.com/greenbone/gsa/pull/1671)
 - Determine the to be applied filter of a list page in GSA and don't rely on the
   backend
@@ -106,6 +117,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   requests in gsad [#1355](https://github.com/greenbone/gsa/pull/1355)
 
 ### Fixed
+- Fixed displaying negative days on the override and note active dashboard [#1727](https://github.com/greenbone/gsa/pull/1727) [#1728](https://github.com/greenbone/gsa/pull/1728)
+- Fixed inability to change to/from LDAP and RADIUS settings [#1723](https://github.com/greenbone/gsa/pull/1723)
 - Fixed filter dialog duplicating filter terms [#1705] (https://github.com/greenbone/gsa/pull/1705)
 - Fixed parsing report details data [#1673](https://github.com/greenbone/gsa/pull/1673)
 - Fixed scanconfig clone icon tooltip does not show if permission is denied [#1664](https://github.com/greenbone/gsa/pull/1664)
@@ -138,6 +151,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   [#1507](https://github.com/greenbone/gsa/pull/1507)
 
 ### Changed
+- Do not simplify filterString in content composer for report download [#1733](https://github.com/greenbone/gsa/pull/1733)
+- Use single component for reloading data [#1722](https://github.com/greenbone/gsa/pull/1722)
+- Use last chars of a label string in BarChart [#1713](https://github.com/greenbone/gsa/pull/1713)
+- Next Scheduled Tasks displays timezone [#1712](https://github.com/greenbone/gsa/pull/1712)
+- Load "small" report before full report [#1697](https://github.com/greenbone/gsa/pull/1697)
 - Removed Clone and Verify functionalities for report formats [#1650](https://github.com/greenbone/gsa/pull/1650)
 - Use new [React context API](https://reactjs.org/docs/context.html#api) [#1637](https://github.com/greenbone/gsa/pull/1637)
 - Update response data parsing in Model classes
@@ -158,6 +176,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   logged in [#1508](https://github.com/greenbone/gsa/pull/1508)
 
 ### Fixed
+- Fixed filtering general command permissions in roles [#1734](https://github.com/greenbone/gsa/pull/1734)
+- Fix getting details in delta report [#1732](https://github.com/greenbone/gsa/pull/1732)
+- Include results (details=1) in report download [#1731](https://github.com/greenbone/gsa/pull/1731)
+- Fixed High column in hosts table fo reports [#1730](https://github.com/greenbone/gsa/pull/1730)
+- Fix setting the user default filter in report results [#1699](https://github.com/greenbone/gsa/pull/1699)
 - Use new model construction in some commands [#1665](https://github.com/greenbone/gsa/pull/1665)
 - Fixed auth_method preselection for user dialog [#1661](https://github.com/greenbone/gsa/pull/1661)
 - Fixed userId, roleId and groupId not rendering in create entity permissions dialog [#1658](https://github.com/greenbone/gsa/pull/1658)